flex-translator-backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
apispec==6.8.2
Authlib==1.6.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
deepl==1.22.0
distro==1.9.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.2
fsspec==2024.10.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
openai==1.52.2
opencv-python-headless==4.11.0.86
packaging==24.1
pdf2docx==0.5.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
PyMuPDF==1.26.1
PyMySQL==1.1.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.13.1
termcolor==3.1.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tqdm==4.66.6
transformers==4.46.0
typing_extensions==4.12.2
urllib3==2.2.3
webargs==8.7.0
Werkzeug==3.0.6
